+++ src/serial/hal_02.rs
@@ -9,7 +9,7 @@ mod nb {
     {
         type Error = Error;
 
-        fn read(&mut self) -> nb::Result<WORD, Error> {
+        fn read(&mut self) -> nb::Result<WORD, Self::Error> {
             self.rx.read()
         }
     }
@@ -18,8 +18,7 @@ mod nb {
         type Error = Error;
 
         fn read(&mut self) -> nb::Result<u8, Self::Error> {
-            // Delegate to the Read<u16> implementation, then truncate to 8 bits
-            Rx::<USART, u16>::new().read().map(|word16| word16 as u8)
+            self.read()
         }
     }
 
@@ -31,33 +30,8 @@ mod nb {
     impl<USART: Instance> Read<u16> for Rx<USART, u16> {
         type Error = Error;
 
-        fn read(&mut self) -> nb::Result<u16, Error> {
-            // NOTE(unsafe) atomic read with no side effects
-            let sr = unsafe { (*USART::ptr()).sr.read() };
-
-            // Any error requires the dr to be read to clear
-            if sr.pe().bit_is_set()
-                || sr.fe().bit_is_set()
-                || sr.nf().bit_is_set()
-                || sr.ore().bit_is_set()
-            {
-                unsafe { (*USART::ptr()).dr.read() };
-            }
-
-            Err(if sr.pe().bit_is_set() {
-                Error::Parity.into()
-            } else if sr.fe().bit_is_set() {
-                Error::Framing.into()
-            } else if sr.nf().bit_is_set() {
-                Error::Noise.into()
-            } else if sr.ore().bit_is_set() {
-                Error::Overrun.into()
-            } else if sr.rxne().bit_is_set() {
-                // NOTE(unsafe) atomic read from stateless register
-                return Ok(unsafe { &*USART::ptr() }.dr.read().dr().bits());
-            } else {
-                nb::Error::WouldBlock
-            })
+        fn read(&mut self) -> nb::Result<u16, Self::Error> {
+            self.read()
         }
     }
 
@@ -81,13 +55,11 @@ mod nb {
         type Error = Error;
 
         fn write(&mut self, word: u8) -> nb::Result<(), Self::Error> {
-            // Delegate to u16 version
-            Tx::<USART, u16>::new().write(u16::from(word))
+            self.write(word)
         }
 
         fn flush(&mut self) -> nb::Result<(), Self::Error> {
-            // Delegate to u16 version
-            Tx::<USART, u16>::new().flush()
+            self.flush()
         }
     }
 
@@ -100,59 +72,28 @@ mod nb {
         type Error = Error;
 
         fn write(&mut self, word: u16) -> nb::Result<(), Self::Error> {
-            // NOTE(unsafe) atomic read with no side effects
-            let sr = unsafe { (*USART::ptr()).sr.read() };
-
-            if sr.txe().bit_is_set() {
-                // NOTE(unsafe) atomic write to stateless register
-                unsafe { &*USART::ptr() }.dr.write(|w| w.dr().bits(word));
-                Ok(())
-            } else {
-                Err(nb::Error::WouldBlock)
-            }
+            self.write(word)
         }
 
         fn flush(&mut self) -> nb::Result<(), Self::Error> {
-            // NOTE(unsafe) atomic read with no side effects
-            let sr = unsafe { (*USART::ptr()).sr.read() };
-
-            if sr.tc().bit_is_set() {
-                Ok(())
-            } else {
-                Err(nb::Error::WouldBlock)
-            }
+            self.flush()
         }
     }
 }
 
 mod blocking {
     use super::super::{Error, Instance, Serial, Tx};
-    use embedded_hal::{blocking::serial::Write, serial};
+    use embedded_hal::blocking::serial::Write;
 
     impl<USART: Instance> Write<u8> for Tx<USART, u8> {
         type Error = Error;
 
         fn bwrite_all(&mut self, bytes: &[u8]) -> Result<(), Self::Error> {
-            for &b in bytes {
-                loop {
-                    match <Self as serial::Write<u8>>::write(self, b) {
-                        Err(nb::Error::WouldBlock) => continue,
-                        Err(nb::Error::Other(err)) => return Err(err),
-                        Ok(()) => break,
-                    }
-                }
-            }
-            Ok(())
+            self.bwrite_all(bytes)
         }
 
         fn bflush(&mut self) -> Result<(), Self::Error> {
-            loop {
-                match <Self as serial::Write<u8>>::flush(self) {
-                    Ok(()) => return Ok(()),
-                    Err(nb::Error::WouldBlock) => continue,
-                    Err(nb::Error::Other(err)) => return Err(err),
-                }
-            }
+            self.bflush()
         }
     }
 
@@ -171,35 +112,20 @@ mod blocking {
     impl<USART: Instance> Write<u16> for Tx<USART, u16> {
         type Error = Error;
 
-        fn bwrite_all(&mut self, buffer: &[u16]) -> Result<(), Self::Error> {
-            for &b in buffer {
-                loop {
-                    match <Self as serial::Write<u16>>::write(self, b) {
-                        Err(nb::Error::WouldBlock) => continue,
-                        Err(nb::Error::Other(err)) => return Err(err),
-                        Ok(()) => break,
-                    }
-                }
-            }
-            Ok(())
+        fn bwrite_all(&mut self, slice: &[u16]) -> Result<(), Self::Error> {
+            self.bwrite_all(slice)
         }
 
         fn bflush(&mut self) -> Result<(), Self::Error> {
-            loop {
-                match <Self as serial::Write<u16>>::flush(self) {
-                    Ok(()) => return Ok(()),
-                    Err(nb::Error::WouldBlock) => continue,
-                    Err(nb::Error::Other(err)) => return Err(err),
-                }
-            }
+            self.bflush()
         }
     }
 
     impl<USART: Instance, PINS> Write<u16> for Serial<USART, PINS, u16> {
         type Error = Error;
 
-        fn bwrite_all(&mut self, bytes: &[u16]) -> Result<(), Self::Error> {
-            self.tx.bwrite_all(bytes)
+        fn bwrite_all(&mut self, slice: &[u16]) -> Result<(), Self::Error> {
+            self.tx.bwrite_all(slice)
         }
 
         fn bflush(&mut self) -> Result<(), Self::Error> {