+++ src/serial/hal_1.rs
@@ -1,15 +1,4 @@
-use embedded_hal_one::serial::{Error, ErrorKind, ErrorType};
-
-impl Error for super::Error {
-    fn kind(&self) -> ErrorKind {
-        match self {
-            Self::Overrun => ErrorKind::Overrun,
-            Self::Framing => ErrorKind::FrameFormat,
-            Self::Parity => ErrorKind::Parity,
-            Self::Noise => ErrorKind::Noise,
-        }
-    }
-}
+use embedded_hal_one::serial::ErrorType;
 
 impl<USART, PINS, WORD> ErrorType for super::Serial<USART, PINS, WORD> {
     type Error = super::Error;
@@ -33,7 +22,7 @@ mod nb {
     impl<USART, PINS, WORD: Copy> Read<WORD> for Serial<USART, PINS, WORD>
     where
         USART: Instance,
-        Rx<USART, WORD>: Read<WORD> + ErrorType<Error = Error>,
+        Rx<USART, WORD>: Read<WORD> + ErrorType<Error = Self::Error>,
     {
         fn read(&mut self) -> nb::Result<WORD, Error> {
             self.rx.read()
@@ -42,8 +31,7 @@ mod nb {
 
     impl<USART: Instance> Read<u8> for Rx<USART, u8> {
         fn read(&mut self) -> nb::Result<u8, Self::Error> {
-            // Delegate to the Read<u16> implementation, then truncate to 8 bits
-            Rx::<USART, u16>::new().read().map(|word16| word16 as u8)
+            self.read()
         }
     }
 
@@ -53,40 +41,15 @@ mod nb {
     /// 9 received data bits and all other bits set to zero. Otherwise, the returned value will contain
     /// 8 received data bits and all other bits set to zero.
     impl<USART: Instance> Read<u16> for Rx<USART, u16> {
-        fn read(&mut self) -> nb::Result<u16, Error> {
-            // NOTE(unsafe) atomic read with no side effects
-            let sr = unsafe { (*USART::ptr()).sr.read() };
-
-            // Any error requires the dr to be read to clear
-            if sr.pe().bit_is_set()
-                || sr.fe().bit_is_set()
-                || sr.nf().bit_is_set()
-                || sr.ore().bit_is_set()
-            {
-                unsafe { (*USART::ptr()).dr.read() };
-            }
-
-            Err(if sr.pe().bit_is_set() {
-                Error::Parity.into()
-            } else if sr.fe().bit_is_set() {
-                Error::Framing.into()
-            } else if sr.nf().bit_is_set() {
-                Error::Noise.into()
-            } else if sr.ore().bit_is_set() {
-                Error::Overrun.into()
-            } else if sr.rxne().bit_is_set() {
-                // NOTE(unsafe) atomic read from stateless register
-                return Ok(unsafe { &*USART::ptr() }.dr.read().dr().bits());
-            } else {
-                nb::Error::WouldBlock
-            })
+        fn read(&mut self) -> nb::Result<u16, Self::Error> {
+            self.read()
         }
     }
 
     impl<USART, PINS, WORD: Copy> Write<WORD> for Serial<USART, PINS, WORD>
     where
         USART: Instance,
-        Tx<USART, WORD>: Write<WORD> + ErrorType<Error = Error>,
+        Tx<USART, WORD>: Write<WORD> + ErrorType<Error = Self::Error>,
     {
         fn flush(&mut self) -> nb::Result<(), Self::Error> {
             self.tx.flush()
@@ -99,13 +62,11 @@ mod nb {
 
     impl<USART: Instance> Write<u8> for Tx<USART, u8> {
         fn write(&mut self, word: u8) -> nb::Result<(), Self::Error> {
-            // Delegate to u16 version
-            Tx::<USART, u16>::new().write(u16::from(word))
+            self.write(word)
         }
 
         fn flush(&mut self) -> nb::Result<(), Self::Error> {
-            // Delegate to u16 version
-            Tx::<USART, u16>::new().flush()
+            self.flush()
         }
     }
 
@@ -116,102 +77,51 @@ mod nb {
     /// will be transmitted and the other 8 bits will be ignored.
     impl<USART: Instance> Write<u16> for Tx<USART, u16> {
         fn write(&mut self, word: u16) -> nb::Result<(), Self::Error> {
-            // NOTE(unsafe) atomic read with no side effects
-            let sr = unsafe { (*USART::ptr()).sr.read() };
-
-            if sr.txe().bit_is_set() {
-                // NOTE(unsafe) atomic write to stateless register
-                unsafe { &*USART::ptr() }.dr.write(|w| w.dr().bits(word));
-                Ok(())
-            } else {
-                Err(nb::Error::WouldBlock)
-            }
+            self.write(word)
         }
 
         fn flush(&mut self) -> nb::Result<(), Self::Error> {
-            // NOTE(unsafe) atomic read with no side effects
-            let sr = unsafe { (*USART::ptr()).sr.read() };
-
-            if sr.tc().bit_is_set() {
-                Ok(())
-            } else {
-                Err(nb::Error::WouldBlock)
-            }
+            self.flush()
         }
     }
 }
 
 mod blocking {
     use super::super::{Instance, Serial, Tx};
-    use embedded_hal_one::serial::{self, blocking::Write};
+    use super::ErrorType;
+    use embedded_hal_one::serial::blocking::Write;
 
-    impl<USART: Instance> Write<u8> for Tx<USART, u8> {
-        fn write(&mut self, bytes: &[u8]) -> Result<(), Self::Error> {
-            for &b in bytes {
-                loop {
-                    match <Self as serial::nb::Write<u8>>::write(self, b) {
-                        Err(nb::Error::WouldBlock) => continue,
-                        Err(nb::Error::Other(err)) => return Err(err),
-                        Ok(()) => break,
-                    }
-                }
-            }
-            Ok(())
-        }
-
-        fn flush(&mut self) -> Result<(), Self::Error> {
-            loop {
-                match <Self as serial::nb::Write<u8>>::flush(self) {
-                    Ok(()) => return Ok(()),
-                    Err(nb::Error::WouldBlock) => continue,
-                    Err(nb::Error::Other(err)) => return Err(err),
-                }
-            }
-        }
-    }
-
-    impl<USART: Instance, PINS> Write<u8> for Serial<USART, PINS, u8> {
-        fn write(&mut self, bytes: &[u8]) -> Result<(), Self::Error> {
-            self.tx.write(bytes)
+    impl<USART, PINS, WORD: Copy> Write<WORD> for Serial<USART, PINS, WORD>
+    where
+        USART: Instance,
+        Tx<USART, WORD>: Write<WORD> + ErrorType<Error = Self::Error>,
+    {
+        fn write(&mut self, slice: &[WORD]) -> Result<(), Self::Error> {
+            self.tx.write(slice)
         }
 
         fn flush(&mut self) -> Result<(), Self::Error> {
             self.tx.flush()
+        }
+    }
+
+    impl<USART: Instance> Write<u8> for Tx<USART, u8> {
+        fn write(&mut self, bytes: &[u8]) -> Result<(), Self::Error> {
+            self.bwrite_all(bytes)
+        }
+
+        fn flush(&mut self) -> Result<(), Self::Error> {
+            self.bflush()
         }
     }
 
     impl<USART: Instance> Write<u16> for Tx<USART, u16> {
-        fn write(&mut self, buffer: &[u16]) -> Result<(), Self::Error> {
-            for &b in buffer {
-                loop {
-                    match <Self as serial::nb::Write<u16>>::write(self, b) {
-                        Err(nb::Error::WouldBlock) => continue,
-                        Err(nb::Error::Other(err)) => return Err(err),
-                        Ok(()) => break,
-                    }
-                }
-            }
-            Ok(())
+        fn write(&mut self, slice: &[u16]) -> Result<(), Self::Error> {
+            self.bwrite_all(slice)
         }
 
         fn flush(&mut self) -> Result<(), Self::Error> {
-            loop {
-                match <Self as serial::nb::Write<u16>>::flush(self) {
-                    Ok(()) => return Ok(()),
-                    Err(nb::Error::WouldBlock) => continue,
-                    Err(nb::Error::Other(err)) => return Err(err),
-                }
-            }
-        }
-    }
-
-    impl<USART: Instance, PINS> Write<u16> for Serial<USART, PINS, u16> {
-        fn write(&mut self, bytes: &[u16]) -> Result<(), Self::Error> {
-            self.tx.write(bytes)
-        }
-
-        fn flush(&mut self) -> Result<(), Self::Error> {
-            self.tx.flush()
+            self.bflush()
         }
     }
 }